pyproject.toml
[tool.poetry]
name = "retrieval-bot"
version = "0.1.0"
description = ""
authors = ["MaGeKo-Exxeta <matti.gerrit.korff@exxeta.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.8"
openai = "^1.12.0"
streamlit = "^1.31.1"
langchain-community = "^0.0.21"
langchain-openai = "^0.0.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
